Created the WavSave class
this class is responsible for the buffer of audio blocks that we are
extracting from the processBlock.

The issue is we can not save to disk reliably in the processBlock
thread.  If there is any hangup at all on the harddrive, it will crash
the DAW.  So we need to offload this save process to a separate thread.

This class is inheriting from juce::Thread and creates a fifo ring
buffer as well.