Started setting up DataStore
So the DataStore class is my attempt at syncing all the instances of the
plugin so that they can all interact with eachother.

This also keeps the save settings in a bit of an easier state.  Before
there was 2 or possibly 3 places that the same variable was being saved
at.  This attempts to simplify this.  Any settings that need to persist
between shutdowns OR need to be shared between instances can go into the
DataStore.

This also helps simplify the code for loading and saving to the plugin
blob data that saves with the DAW.