Updated DataStore
Added checks for if we try to call data that doesnt exist,

also revamped the way that data is called and removed.  Because we can
not use a vector for these objects (juce doesnt allow it), then we need
to find the newest available slot in the Versions array to add our new
instance data to.  This update streamlines all of that for us.