Issues with TabInstance
This commit is to just document my latest point in this.  Some issues
have been coming up and I'm going to revert back a bit.

The issue with the Table component is the custom components i can put
inside the cells.  I am attempting to update the textboxes automatically
and the table object does not like it very much..

Going to try a different approach.