Successfully created our own TextEditor class
So we will need to create custom editor classes for each of the values
we want to control.  This allows dynamic updating of variables as we
type and keeps all of the instances in sync.

The only other issue I need to test is if after we create our own
TextEditor class, will it still be considered a component?  Can we still
use these in a Table class?