Finally have realtime updates for Table object.. sortof
So, we now have text boxes that update the back-end values in realtime.
Just need to create a custom class per text box.

Only other thing that is becoming an issue is the updating of other
components in real-time.  Currently to test, I have two textboxes
assigned to the same value in the table.  When I type in one, the other
one should update in realtime.. Not the case.  I have to click on the
other textbox and THEN it will update to the new value.

This is because I am currently overriding the paint function and
apparently this function is not being called very frequently to save
processing power.

Need to figure a way around this.