SUCCESSFUL!!
OK finally got a working version of this!

Each individual instance can record individually or as a group.  If the
group name is the same, the record button will trigger everything in
that group.

This also is mappable to midi and to key maps in Ableton!

Still has a few bugs:
- Record buttons (or any buttons for that matter) do not sync quickly,
  you have to run the mouse over the buttons for them to update.
- Make it more obvious which line is the current instance.  can be
  slightly confusing.
- Still need to make the rec on play switch work.
- Update the check boxes in the same group
- Figure out some way of updating all the group textboxes at once. Issue
  is you still want to be able to take an instance out of a group.. hmm.