Fixed issue with delete then add plugin instance
There was an issue where if we deleted an instance and added another
one, the DAW would crash.  The table was looking up instances by row
number and not ID number so we were looking for a position in the insts
array that didnt exist.

This adds a new function to the DataStore that looks up the ID number
from a specified position in the insts array.