Got a working version of ValueTree working
Using TextBox.h, we now have a working class we can use for any textbox
syncing we need.  just give it a texteditor id name, the instance name
to control, and the variable name in the valuetree to sync to.  This
will do the rest!

We should easily be able to do the same thing for buttons and
checkboxes.